Killed the Moose. Chopped an x::Class off of the DBI
This was really more complex than it needed to be, and it really wasn't
as lightweight as I wanted it to be. I removed Moose, as well as the
DBIx::Class approach, opting for straight DBI and a more script-oriented
approach. It's much lighter now, perhaps at the expense of some
cuteness.